scrap/requirements_db.txt
psycopg2-binary==2.9.9
tqdm==4.66.1

